Extract autolinking logic out of react-native-podspecs into react-native-autolinking (as TS)
