+++ packages/react-native-autolinking/README.md
@@ -1,12 +1,23 @@
 # @ammarahm-ed/react-native-autolinking
 
-```javascript
-ns plugin add @ammarahm-ed/react-native-autolinking
+A library of autolinking utilities to be called by other packages (usually upon a NativeScript lifecycle hook).
+
+```sh
+npm install --save @ammarahm-ed/react-native-autolinking
 ```
 
 ## Usage
 
-// TODO
+```ts
+import { autolinkIos } from '@ammarahm-ed/react-native-autolinking';
+
+// This is how you'd call it, though see '@ammarahm-ed/react-native-podspecs'
+// for how to use it in the context of a hook.
+await autolinkIos({
+  dependencies: ['@ammarahm-ed/react-native-module-test'],
+  projectDir: '/Users/jamie/Documents/git/nativescript-magic-spells/apps/demo',
+});
+```
 
 ## License
