+++ packages/react-native-autolinking/index.ts
@@ -1,1 +1,2 @@
 export { autolinkIos } from './autolinkIos';
+export type { HookArgs } from './HookArgs';
