+++ packages/react-native-podspecs/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
+    "module": "CommonJS",
     "outDir": "../../dist/out-tsc",
-    "rootDir": "."
+    "rootDirs": ["."]
   },
   "exclude": ["**/*.spec.ts", "**/*.test.ts", "angular"],
   "include": ["**/*.ts", "references.d.ts"]
